+++ src/tools/build-manifest/src/main.rs
@@ -149,7 +149,7 @@ static TARGETS: &[&str] = &[
     "wasm32-unknown-unknown",
     "wasm32-wasi",
     "wasm32-wasip1",
-    "wasm32-wasi-preview1-threads",
+    "wasm32-wasip1-threads",
     "x86_64-apple-darwin",
     "x86_64-apple-ios",
     "x86_64-fortanix-unknown-sgx",
